Refactor resume extraction logic and remove unused database connection code

+++ job_blueprint/main.py
@@ -2,10 +2,10 @@ from fastapi import FastAPI, File, HTTPException, UploadFile, Form
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import StreamingResponse
 from agent import chat
-import json
 import os
 from libs.extraction import spacy_extraction, llm_extraction
 from libs.convect import pdf_to_markdown , pdf_to_markdown_EasyOCR , pdf_to_markdown_MistralOCR , any_to_markdown
+from database.handler import import_skil
 app = FastAPI()
 
 app.add_middleware(
@@ -34,7 +34,7 @@ async def chat_stream():
     return StreamingResponse(fake_stream(), media_type="text/plain")
 
 
-
+ 
 @app.post("/extract")
 async def extract_text(
     file: UploadFile = File(...),
@@ -46,9 +46,17 @@ async def extract_text(
     filename  = file.filename
     save_path = os.path.join(SAVE_DIR, filename)
 
+    if os.path.exists(save_path):
+        raise HTTPException(
+            status_code=400,
+            detail=f"ไฟล์ '{filename}' มีอยู่แล้วในระบบ"
+        )
+
+
     content = await file.read()
     with open(save_path, "wb") as f:
         f.write(content)
+
 
     if read_by == "default" and filename.endswith(".pdf"):
         read_by = "pdf"
@@ -62,30 +70,53 @@ async def extract_text(
             markdown = pdf_to_markdown_MistralOCR(save_path)
         case "easy":
             markdown = pdf_to_markdown_EasyOCR(content)
+        case _:
+            raise HTTPException(status_code=400, detail="Invalid OCR type. Choose: default, easy, mistral")
+        
     if len(markdown) < 100:
         raise HTTPException(
             status_code=400,
             detail="ไม่สามารถสกัดเอกสารได้ ลองใช้เป็น `mistral`"
         )
 
-    ner = spacy_extraction(markdown)
-    return {"ner": ner}
+    try:
+        ner = spacy_extraction(markdown)
+        blueprint = llm_extraction(ner, markdown, "gpt-4.1-mini")
 
+        status = import_skil(blueprint)
 
+        return status
+
+    except Exception as e:
+        raise HTTPException(
+            status_code=400,
+            detail=str(e)
+        )
     
-    # except Exception as e:
-    #     return {"error": str(e)}
+    
+# from elasticsearch import Elasticsearch
 
-    # content = await file.read()
+# es = Elasticsearch("http://localhost:9200")
 
+# query = {
+#     "size": 0,
+#     "query": {
+#         "match_phrase": {
+#             "position": "Fullstack Developer"
+#         }
+#     },
+#     "aggs": {
+#         "popular_skills": {
+#             "terms": {
+#                 "field": "skill.keyword",
+#                 "size": 10
+#             }
+#         }
+#     }
+# }
 
-    # if ocr_type == "default":
-    #     markdown = file_to_markdown(content)
-    # elif ocr_type == "easy":
-    #     markdown = pdf_to_markdown_EasyOCR(content)
-    # elif ocr_type == "mistral":
-    #     markdown = pdf_to_markdown_MistralOCR(content)
-    # else:
-    #     raise HTTPException(status_code=400, detail="Invalid OCR type. Choose: default, easy, mistral")
+# response = es.search(index="skills_by_position", body=query)
 
-    # return PlainTextResponse(markdown)
+# # แสดงผลสกิลที่นิยมที่สุด
+# for bucket in response['aggregations']['popular_skills']['buckets']:
+#     print(f"Skill: {bucket['key']}, Count: {bucket['doc_count']}")
